requirements.txt
ray[default] 
numpy
scipy
pandas 
absl-py 
Pillow
scikit-learn
opencv-python
guided-filter-pytorch